Projects/Middleware/AuthenticationAuthorization/pyproject.toml
[project]
name = "authenticationauthorization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pyjwt>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
